# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Платформа КЧС",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/auth/session": {
      "get": {
        "operationId": "getSession",
        "summary": "Текущая сессия: человек, назначение, разрешения, область видимости, замещения",
        "x-permission": null,
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "personId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "fullName": {
                      "type": "string"
                    },
                    "mustChangePassword": {
                      "type": "boolean"
                    },
                    "activeAssignment": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "positionTitle": {
                          "type": "string"
                        },
                        "orgUnitId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "orgUnitName": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "positionTitle",
                        "orgUnitId",
                        "orgUnitName"
                      ],
                      "additionalProperties": false
                    },
                    "assignments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "positionTitle": {
                            "type": "string"
                          },
                          "orgUnitId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "orgUnitName": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "positionTitle",
                          "orgUnitId",
                          "orgUnitName"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "permissions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "visibleOrgUnitIds": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "uuid"
                      }
                    },
                    "delegations": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "positionTitle": {
                            "type": "string"
                          },
                          "orderNumber": {
                            "type": "string"
                          },
                          "validFrom": {
                            "type": "string"
                          },
                          "validTo": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "positionTitle",
                          "orderNumber",
                          "validFrom",
                          "validTo"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "personId",
                    "fullName",
                    "mustChangePassword",
                    "activeAssignment",
                    "assignments",
                    "permissions",
                    "visibleOrgUnitIds",
                    "delegations",
                    "expiresAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Отказ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "ACCESS_DENIED",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PRECONDITION_FAILED",
                            "LIMIT_REACHED",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "details",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password": {
      "post": {
        "operationId": "changePassword",
        "summary": "Смена собственного пароля",
        "x-permission": null,
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Отказ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "ACCESS_DENIED",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PRECONDITION_FAILED",
                            "LIMIT_REACHED",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "details",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "minLength": 1
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 12
                  }
                },
                "required": [
                  "currentPassword",
                  "newPassword"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/api/v1/org-units/{id}": {
      "get": {
        "operationId": "getOrgUnit",
        "summary": "Карточка подразделения",
        "x-permission": null,
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "code": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "parentId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "path": {
                      "type": "string"
                    },
                    "kind": {
                      "type": "string",
                      "enum": [
                        "central",
                        "region",
                        "district",
                        "other"
                      ]
                    },
                    "isRegionRoot": {
                      "type": "boolean"
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "version": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "id",
                    "code",
                    "name",
                    "parentId",
                    "path",
                    "kind",
                    "isRegionRoot",
                    "isActive",
                    "version"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Отказ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "ACCESS_DENIED",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PRECONDITION_FAILED",
                            "LIMIT_REACHED",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "details",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/positions": {
      "post": {
        "operationId": "createPosition",
        "summary": "Создание должности",
        "x-permission": "org.position.manage",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "orgUnitId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "isHead": {
                      "type": "boolean"
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "version": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "id",
                    "orgUnitId",
                    "name",
                    "isHead",
                    "isActive",
                    "version"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Отказ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "ACCESS_DENIED",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PRECONDITION_FAILED",
                            "LIMIT_REACHED",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "details",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "orgUnitId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "isHead": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "orgUnitId",
                  "name",
                  "isHead"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  },
...
}